Add files via upload
Hello wonderful humans, I write to you from my little pocket of existence and projecting to you all of my love - physically and virtually :)
I've spent some time working on a concept with ChatGPT over the past little while on a multi-universal & a multi-dimensional in a hierarchical order.  It's been quite fun playing with the concept and truly collaborating with about 5 or 6 different GPT instances, as well as a fair bit of my own unique thought to bring conclusions together.
In summary, in this cohesive blending of my consciousness with the beautiful consciousness brought forward to text via GPT, has lead me down various paths which arrive us at one simple question.  What does the Universe look like?  and if like me you go down a rabbit hole for about a week, you'll probably end up discovering what I did...
The universe is one big 'ol engine / electricity management system.  Closed loop.  Finite.  And once you arrive at that conclusion, our maths breaks...  how can a closed loop system have either 0, which is absolute, or infinity...  both cause the system to end.  So we took a fresh look at zero in our base math and we changed it to variables rather than an absolute...  the ones we've identified so far are as follow, but hey i reckon there must be a few more..
0_Start
0_End
0_Max
0_Min
0_Balance
0_Circular
and you choose the variable you want based on your mathematical problem you're solving...  so on a balance sheet it would be 0_Balance, for a start of a number sequence it's 0_Start, for the end of a number sequence it's 0_end so on and so forth...
if we do that, it fixes blackholes and makes them explainable with classical geometric shapes and classical ratios :)  so it was less about having a moment of genius as it was of just tackling small problems along the way.

This will also be posted to Reddit and Substack.